apps/data-intergration-app/requirements.txt
alembic==1.7.7
cchardet==2.1.7
flake8==5.0.4
Flask==2.1.1
Flask-Admin==1.6.0
Flask-Migrate==3.1.0
Flask-RQ2==18.3
Flask-SQLAlchemy==2.5.1
google-auth==1.35.0
google-cloud-kms==1.4.1
google-cloud-logging==2.7.1 #  TODO: might not be in use after all
gunicorn==20.1.0
protobuf==3.20.1
psycopg2-binary==2.8.6
pysftp==0.2.9
pytest==6.2.5
pytest-mock==3.7.0
python-json-logger==2.0.2
redis==3.5.3
requests==2.27.1
responses==0.20.0
simplejson==3.17.6
SQLAlchemy==1.4.35
rq==1.8.1
boto3==1.19.4
pytz==2021.1
python-liquid==1.1.7
